+++ entrypoint.sh
@@ -13,7 +13,7 @@ function print_help {
 
 case ${1} in
     start)
-        exec python -m rasa_core.server "${@:2}"
+        exec python -m rasa_core.run --enable_api "${@:2}"
         ;;
     run)
         exec "${@:2}"