+++ rasa_core/run.py
@@ -6,6 +6,9 @@ from __future__ import unicode_literals
 import argparse
 import logging
 from collections import namedtuple
+
+from flask import Flask
+from flask_cors import CORS
 from threading import Thread
 
 from builtins import str
@@ -228,12 +231,17 @@ def start_server(input_channels,
                  cors,
                  auth_token,
                  port,
-                 initial_agent):
+                 initial_agent,
+                 enable_api=True):
     """Run the agent."""
 
-    app = server.create_app(initial_agent,
-                            cors_origins=cors,
-                            auth_token=auth_token)
+    if enable_api:
+        app = server.create_app(initial_agent,
+                                cors_origins=cors,
+                                auth_token=auth_token)
+    else:
+        app = Flask(__name__)
+        CORS(app, resources={r"/*": {"origins": cors or ""}})
 
     if input_channels:
         rasa_core.channels.channel.register(input_channels,
@@ -256,12 +264,13 @@ def serve_application(initial_agent,
                       port=constants.DEFAULT_SERVER_PORT,
                       credentials_file=None,
                       cors=None,
-                      auth_token=None
+                      auth_token=None,
+                        enable_api=True
                       ):
     input_channels = create_http_input_channels(channel, credentials_file)
 
     http_server = start_server(input_channels, cors, auth_token,
-                               port, initial_agent)
+                               port, initial_agent, enable_api)
 
     if channel == "cmdline":
         start_cmdline_io(constants.DEFAULT_SERVER_URL, http_server.stop)
@@ -318,4 +327,5 @@ if __name__ == '__main__':
                       cmdline_args.port,
                       cmdline_args.credentials,
                       cmdline_args.cors,
-                      cmdline_args.auth_token)
+                      cmdline_args.auth_token,
+                      cmdline_args.enable_api)